requirements.txt
tqdm==4.43.0
pandas==0.25.1
numpy==1.17.2
matplotlib==3.1.1
pyvi==0.0.9.7
sklearn-crfsuite==0.3.6
